+++ src/Microsoft.Identity.Web/MergedOptions.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.Identity.Client;
@@ -186,10 +187,7 @@ namespace Microsoft.Identity.Web
                 mergedOptions.ResponseMode = microsoftIdentityOptions.ResponseMode;
             }
 
-            if (string.IsNullOrEmpty(mergedOptions.ResponseType) && !string.IsNullOrEmpty(microsoftIdentityOptions.ResponseType))
-            {
-                mergedOptions.ResponseType = microsoftIdentityOptions.ResponseType;
-            }
+            mergedOptions.ResponseType = microsoftIdentityOptions.ResponseType;
 
             if (string.IsNullOrEmpty(mergedOptions.ReturnUrlParameter) && !string.IsNullOrEmpty(microsoftIdentityOptions.ReturnUrlParameter))
             {
@@ -352,9 +350,9 @@ namespace Microsoft.Identity.Web
 
         public void PrepareAuthorityInstanceForMsal()
         {
-            if (IsB2C && Instance.EndsWith("/tfp/"))
+            if (IsB2C && Instance.EndsWith("/tfp/", StringComparison.OrdinalIgnoreCase))
             {
-                Instance = Instance.Replace("/tfp/", string.Empty).TrimEnd('/') + "/";
+                Instance = Instance.Replace("/tfp/", string.Empty, StringComparison.OrdinalIgnoreCase).TrimEnd('/') + "/";
             }
             else
             {