+++ tests/WebAppCallsWebApiCallsGraph/TodoListService/Controllers/TodoListController.cs
@@ -15,7 +15,7 @@ namespace TodoListService.Controllers
 {
     [Authorize]
     [Route("api/[controller]")]
-    [RequiredScope("access_as_user")] 
+    [RequiredScope(RequiredScopesConfigurationKey = "AzureAd:Scopes")]
     public class TodoListController : Controller
     {
         private readonly ITokenAcquisition _tokenAcquisition; // do not remove